refactor: move preset loading upwards
