+++ src/App.css
@@ -1,3 +1,7 @@
+html, body, #root {
+  height: 100%;
+}
+
 #root {
   width: 100vw;
   margin: 0 auto;
@@ -10,14 +14,20 @@
   margin-top: 64px;
 }
 
+.App--loading {
+  /* TODO */
+}
+
 .logo {
   height: 6em;
   padding: 1.5em;
   will-change: filter;
 }
+
 .logo:hover {
   filter: drop-shadow(0 0 2em #646cffaa);
 }
+
 .logo.react:hover {
   filter: drop-shadow(0 0 2em #61dafbaa);
 }
@@ -62,6 +72,10 @@
   width: 100%;
 }
 
+.height-100 {
+  height: 100%;
+}
+
 .d-flex {
   display: flex;
 }
@@ -81,6 +95,10 @@
 
 .m-8 {
   margin: 8px;
+}
+
+.m-16 {
+  margin: 16px;
 }
 
 .mh-8 {
@@ -138,17 +156,5 @@
 
   .width-50 {
     width: 100%;
-  }
-}
-
-@media (max-width: 400px) {
-  .App {
-    /* margin-top: 140px; */
-  }
-}
-
-@media (max-width: 308px) {
-  .App {
-    /* margin-top: 180px; */
   }
 }
